backend/requirements.txt
db-dtypes>=1.0.0
google-cloud-bigquery>=3.0.0
pandas>=1.5.0
pyarrow>=10.0.0
pymilvus>=2.3.0,<2.4.0
openai==0.28.1
cohere>=4.0.0
python-dotenv>=0.21.0
uvicorn>=0.20.0
fastapi>=0.95.0
httpx>=0.23.0
pydantic>=1.10.0
scikit-learn>=1.0.0
tenacity>=8.0.0
llama-index==0.8.4
langchain==0.0.266
langchain-community>=0.0.1
pymupdf>=1.20.0
numpy>=1.20.0